requirements.txt
Django==2.1.1
psycopg2
django-model-utils
Pillow
django-rest-framework1
django-ckeditor

